pyproject.toml
[tool.poetry]
name = "xbrl"
version = "0.1.0"
description = ""
authors = ["lynma01 <lynma01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
py-xbrl = "^2.2.6"
httpx = "^0.23.3"
beautifulsoup4 = "^4.12.0"
lxml = "^4.9.2"
polars = "^0.16.18"
duckdb = "^0.7.1"
jupysql = "^0.6.6"
pyarrow = "^11.0.0"
pytest = "^7.2.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
